requirements.txt
langchain==0.1.0
langchain-community==0.0.20
langchain-openai==0.0.5
openai==1.12.0
chromadb==0.4.22
sentence-transformers==2.2.2
pypdf==3.17.1
python-dotenv==1.0.0
fastapi==0.104.1
uvicorn==0.24.0
groq==0.4.1
